requirements.txt
#only include the libraries used in the project
pdfplumber
pandas
numpy
openpyxl
google-generativeai
python-dotenv
configparser
linkedin-api